Add ReadIdleTimeout to http2 transport
This fixes a bug scenario in which during certain OS network disturbances
existing http2 connections fail to receive data, but will not be closed.

Previous to this fix, we would continue to write data to that connection
and end up in a stuck/permanently idle state.

From the ReadIdleTimeout docs:

// ReadIdleTimeout is the timeout after which a health check using ping
// frame will be carried out if no frame is received on the connection.
// Note that a ping response will is considered a received frame, so if
// there is no other traffic on the connection, the health check will
// be performed every ReadIdleTimeout interval.
// If zero, no health check is performed.

By setting a ReadIdleTimeout, in those scenarios in which existing http2
connections are no longer receiving data, after N time it will attempt a
health check by sending a PING frame and if no response is received it
will close/cleanup that connection (allowing us to retry/create a new
one that can be successful).